ac/llvm: implement WA in nir to llvm
LLVM has the FeatureVALUTransUseHazard workaround for gfx11.
The problem is that this doesn't work for shader built in
parts.

LLVM will be modified to be more conservative and apply the
workaround in more places but in the meantime, add a simpler
implementation in the nir to llvm backend.

The workaround is only added is transcendental VALU ops are
used which should cover most cases where the WA is needed,
withtout penalizing shaders that don't require it.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/10785
Cc: mesa-stable