anv/sparse: add struct anv_sparse_surf_info
In the sparse code we often have to find out the isl_surf related to
the surfaces we're dealing with, and then we compute/use a bunch of
properties of that surface. The problem is that this is done in quite
a few different places, so the code to "extract sparse-relevant
information from struct isl_surf" is duplicated in many places. This
creates two problems:

  - If the computation of these derived values is wrong, when we fix
    them we'll need to find all the places where this computation
    happens and fix them. We recently had to fix tile size computation
    in 3 different places when we added multi-sampling support to
    sparse.

  - Computation of some of these values may be somewhat expensive. For
    example, isl_surf_get_tile_info() is not trivial and it is called
    multiple times during anv_image_get_sparse_memory_requirements()
    subfunctions.

So in this commit we add struct anv_sparse_surf_info and, for now, we
just add struct isl_surf to it and then change the anv_sparse
functions to pass this struct instead of struct isl_surf everywhere.
The next commits will take care of expanding this struct with more
useful data.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>