mesa/main: remove non-existent limitations
The OpenGL 4.6 compatibility spec does not have any of these limiations,
so we should really allow this. And we have working fall-back code for
all formats, so there's really no reason we shouldn't allow it.

But let's keep the ASTC limitation. That is introduced by the
KHR_texture_compression_astc_[hl]dr extensions, which written against
the OpenGL ES 3.1 spec. And the ES3 spec does limit glGenerateMipmap to
color-renderable formats, which these aren't.